+++ lib/src/scanner_text_input.dart
@@ -1,0 +1,2 @@
+export 'scanner_text_input_stub.dart'
+    if (dart.library.io) 'scanner_text_input_io.dart';