+++ lib/widget_utilities.dart
@@ -1,2 +1,3 @@
 export 'src/refresh_universal.dart';
-export 'src/date_piker_universal.dart';
+export 'src/date_picker_universal.dart';
+export 'src/scanner_text_input.dart';